backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
pydantic
pydantic[email]
python-multipart
cryptography